refactor(splash-page): update data import paths for improved modularity
- Changed imports from "@/lib/homepage-data" to "@/lib/splash-page-data" across multiple components.
- This adjustment enhances the modularity of the splash page components, aligning with the new data structure for better organization and maintainability.
