+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
 import "./globals.css";
 import { PHProvider, ThemeProvider } from "@/providers";
-import { SITE_METADATA } from "@/lib/homepage-data";
+import { SITE_METADATA } from "@/lib/splash-page-data";
 import { Toaster } from "sonner";
 
 const poppins = Poppins({
